+++ doc/source/doxygen-docs/changelog.md
@@ -6,6 +6,7 @@
     - Useless nanogui_win() converted into mrpt::gui::CDisplayWindowGUI::nanogui_screen()
     - nanogui: New methods: nanogui::Screen::mouseState(), nanogui::Screen::mouseModifiers()
 - BUG FIXES:
+  - navlog-viewer: Crash when clicking "play" (Closes #1103)
   - RawLogViewer: Fix wrong indices in tree view. Fix freezed progress bar loading a second rawlog.
   - RawLogViewer: Fix wrong rendering if font is missing in the system (Ubuntu 20.04)
   - rawlog-edit: Fix --cut operation leaving empty sensory frames/action collections.